AMQP cleanup and race condition fixes.
