+++ docs/source/config/inputs/amqp.rst
@@ -8,42 +8,42 @@ needs to be specified in the plugin configuration.
 
 Config:
 
-- URL (string):
+- url (string):
     An AMQP connection string formatted per the `RabbitMQ URI Spec
     <http://www.rabbitmq.com/uri-spec.html>`_.
-- Exchange (string):
+- exchange (string):
     AMQP exchange name
-- ExchangeType (string):
+- exchange_type (string):
     AMQP exchange type (`fanout`, `direct`, `topic`, or `headers`).
-- ExchangeDurability (bool):
+- exchange_durability (bool):
     Whether the exchange should be configured as a durable exchange. Defaults
     to non-durable.
-- ExchangeAutoDelete (bool):
+- exchange_auto_delete (bool):
     Whether the exchange is deleted when all queues have finished and there
     is no publishing. Defaults to auto-delete.
-- RoutingKey (string):
+- routing_key (string):
     The message routing key used to bind the queue to the exchange. Defaults
     to empty string.
-- PrefetchCount (int):
+- prefetch_count (int):
     How many messages to fetch at once before message acks are sent. See
     `RabbitMQ performance measurements
     <http://www.rabbitmq.com/blog/2012/04/25/rabbitmq-performance-
     measurements-part-2/>`_ for help in tuning this number. Defaults to 2.
-- Queue (string):
+- queue (string):
     Name of the queue to consume from, an empty string will have the broker
     generate a name for the queue. Defaults to empty string.
-- QueueDurability (bool):
+- queue_durability (bool):
     Whether the queue is durable or not. Defaults to non-durable.
-- QueueExclusive (bool):
+- queue_exclusive (bool):
     Whether the queue is exclusive (only one consumer allowed) or not.
     Defaults to non-exclusive.
-- QueueAutoDelete (bool):
+- queue_auto_delete (bool):
     Whether the queue is deleted when the last consumer un-subscribes.
     Defaults to auto-delete.
-- QueueTTL (int):
+- queue_ttl (int):
     Allows ability to specify TTL in milliseconds on Queue declaration for
     expiring messages. Defaults to undefined/infinite.
-- Decoder (string):
+- decoder (string):
     Decoder name used to transform a raw message body into a structured hekad
     message. Must be a decoder appropriate for the messages that come in from
     the exchange. If accepting messages that have been generated by an
@@ -65,7 +65,7 @@ consume serialized messages would look like:
     [AMQPInput]
     url = "amqp://guest:guest@rabbitmq/"
     exchange = "testout"
-    exchangeType = "fanout"
+    exchange_type = "fanout"
 
 Or you might use a PayloadRegexDecoder to parse OSX syslog messages with the
 following:
@@ -75,7 +75,7 @@ following:
     [AMQPInput]
     url = "amqp://guest:guest@rabbitmq/"
     exchange = "testout"
-    exchangeType = "fanout"
+    exchange_type = "fanout"
     decoder = "logparser"
 
     [logparser]
