+++ plugins/amqp/amqp_test.go
@@ -58,12 +58,8 @@ func AMQPPluginSpec(c gs.Context) {
 	// Setup the mock amqpHub with the mock chan return
 	aqh := NewMockAMQPConnectionHub(ctrl)
 	aqh.EXPECT().GetChannel("", AMQPDialer{}).Return(mch, ug, cg, nil)
-	var oldHub AMQPConnectionHub
-	oldHub = amqpHub
-	amqpHub = aqh
-	defer func() {
-		amqpHub = oldHub
-	}()
+
+	errChan := make(chan error, 1)
 
 	c.Specify("An amqp input", func() {
 		// Setup all the mock calls for Init
@@ -87,17 +83,17 @@ func AMQPPluginSpec(c gs.Context) {
 
 		ith.MockInputRunner.EXPECT().InChan().Return(ith.PackSupply)
 
-		msgChan := make(chan *message.Message, 1)
+		amqpInput := new(AMQPInput)
+		amqpInput.amqpHub = aqh
+		config := amqpInput.ConfigStruct().(*AMQPInputConfig)
+		config.URL = ""
+		config.Exchange = ""
+		config.ExchangeType = ""
+		config.RoutingKey = "test"
+		config.QueueTTL = 300000
 
 		c.Specify("with a valid setup and no decoder", func() {
-			amqpInput := new(AMQPInput)
-			defaultConfig := amqpInput.ConfigStruct().(*AMQPInputConfig)
-			defaultConfig.URL = ""
-			defaultConfig.Exchange = ""
-			defaultConfig.ExchangeType = ""
-			defaultConfig.RoutingKey = "test"
-			defaultConfig.QueueTTL = 300000
-			err := amqpInput.Init(defaultConfig)
+			err := amqpInput.Init(config)
 			c.Assume(err, gs.IsNil)
 			c.Expect(amqpInput.ch, gs.Equals, mch)
 
@@ -125,28 +121,22 @@ func AMQPPluginSpec(c gs.Context) {
 
 				ith.PackSupply <- ith.Pack
 				go func() {
-					amqpInput.Run(ith.MockInputRunner, ith.MockHelper)
-					msgChan <- ith.Pack.Message
+					err := amqpInput.Run(ith.MockInputRunner, ith.MockHelper)
+					errChan <- err
 				}()
 				ith.PackSupply <- ith.Pack
 				close(streamChan)
-				msg := <-msgChan
-				c.Expect(msg.GetType(), gs.Equals, "amqp")
-				c.Expect(msg.GetPayload(), gs.Equals, "This is a message")
+				err = <-errChan
+				c.Expect(err, gs.IsNil)
+				c.Expect(ith.Pack.Message.GetType(), gs.Equals, "amqp")
+				c.Expect(ith.Pack.Message.GetPayload(), gs.Equals, "This is a message")
 			})
 		})
 
 		c.Specify("with a valid setup using a decoder", func() {
 			decoderName := "defaultDecoder"
-			amqpInput := new(AMQPInput)
-			defaultConfig := amqpInput.ConfigStruct().(*AMQPInputConfig)
-			defaultConfig.URL = ""
-			defaultConfig.Exchange = ""
-			defaultConfig.ExchangeType = ""
-			defaultConfig.RoutingKey = "test"
-			defaultConfig.Decoder = decoderName
-			defaultConfig.QueueTTL = 300000
-			err := amqpInput.Init(defaultConfig)
+			config.Decoder = decoderName
+			err := amqpInput.Init(config)
 			c.Assume(err, gs.IsNil)
 			c.Expect(amqpInput.ch, gs.Equals, mch)
 
@@ -183,14 +173,14 @@ func AMQPPluginSpec(c gs.Context) {
 
 				ith.PackSupply <- ith.Pack
 				go func() {
-					amqpInput.Run(ith.MockInputRunner, ith.MockHelper)
-					msgChan <- ith.Pack.Message
+					err := amqpInput.Run(ith.MockInputRunner, ith.MockHelper)
+					errChan <- err
 				}()
 				ith.PackSupply <- ith.Pack
 				close(streamChan)
-				msg := <-msgChan
-				c.Expect(msg.GetType(), gs.Equals, "amqp")
-				c.Expect(msg.GetPayload(), gs.Equals, "This is a message")
+				err = <-errChan
+				c.Expect(ith.Pack.Message.GetType(), gs.Equals, "amqp")
+				c.Expect(ith.Pack.Message.GetPayload(), gs.Equals, "This is a message")
 			})
 
 			c.Specify("consumes a serialized message", func() {
@@ -231,7 +221,8 @@ func AMQPPluginSpec(c gs.Context) {
 
 				ith.PackSupply <- ith.Pack
 				go func() {
-					amqpInput.Run(ith.MockInputRunner, ith.MockHelper)
+					err := amqpInput.Run(ith.MockInputRunner, ith.MockHelper)
+					errChan <- err
 				}()
 				packRef := <-ith.DecodeChan
 				c.Expect(ith.Pack, gs.Equals, packRef)
@@ -239,6 +230,8 @@ func AMQPPluginSpec(c gs.Context) {
 				c.Expect(string(packRef.MsgBytes), gs.Equals, string(msgBody[5:]))
 				ith.PackSupply <- ith.Pack
 				close(streamChan)
+				err = <-errChan
+				c.Expect(err, gs.IsNil)
 			})
 		})
 	})
@@ -248,11 +241,12 @@ func AMQPPluginSpec(c gs.Context) {
 		pConfig := NewPipelineConfig(nil)
 
 		amqpOutput := new(AMQPOutput)
-		defaultConfig := amqpOutput.ConfigStruct().(*AMQPOutputConfig)
-		defaultConfig.URL = ""
-		defaultConfig.Exchange = ""
-		defaultConfig.ExchangeType = ""
-		defaultConfig.RoutingKey = "test"
+		amqpOutput.amqpHub = aqh
+		config := amqpOutput.ConfigStruct().(*AMQPOutputConfig)
+		config.URL = ""
+		config.Exchange = ""
+		config.ExchangeType = ""
+		config.RoutingKey = "test"
 
 		closeChan := make(chan *amqp.Error)
 
@@ -281,12 +275,12 @@ func AMQPPluginSpec(c gs.Context) {
 			encoder.Init(econfig)
 			payloadBytes, err := encoder.Encode(pack)
 
-			defaultConfig.Encoder = "PayloadEncoder"
-			defaultConfig.ContentType = "text/plain"
+			config.Encoder = "PayloadEncoder"
+			config.ContentType = "text/plain"
 			oth.MockOutputRunner.EXPECT().Encoder().Return(encoder)
 			oth.MockOutputRunner.EXPECT().Encode(pack).Return(payloadBytes, nil)
 
-			err = amqpOutput.Init(defaultConfig)
+			err = amqpOutput.Init(config)
 			c.Assume(err, gs.IsNil)
 			c.Expect(amqpOutput.ch, gs.Equals, mch)
 
@@ -295,11 +289,13 @@ func AMQPPluginSpec(c gs.Context) {
 			close(inChan)
 
 			go func() {
-				amqpOutput.Run(oth.MockOutputRunner, oth.MockHelper)
+				err := amqpOutput.Run(oth.MockOutputRunner, oth.MockHelper)
+				errChan <- err
 			}()
 
 			ug.Wait()
-
+			err = <-errChan
+			c.Expect(err, gs.IsNil)
 		})
 
 		c.Specify("publishes a serialized message", func() {
@@ -310,7 +306,7 @@ func AMQPPluginSpec(c gs.Context) {
 			oth.MockOutputRunner.EXPECT().Encoder().Return(encoder)
 			oth.MockOutputRunner.EXPECT().Encode(pack).Return(protoBytes, nil)
 
-			err = amqpOutput.Init(defaultConfig)
+			err = amqpOutput.Init(config)
 			c.Assume(err, gs.IsNil)
 			c.Expect(amqpOutput.ch, gs.Equals, mch)
 
@@ -319,9 +315,12 @@ func AMQPPluginSpec(c gs.Context) {
 			close(inChan)
 
 			go func() {
-				amqpOutput.Run(oth.MockOutputRunner, oth.MockHelper)
+				err := amqpOutput.Run(oth.MockOutputRunner, oth.MockHelper)
+				errChan <- err
 			}()
 			ug.Wait()
+			err = <-errChan
+			c.Expect(err, gs.IsNil)
 		})
 	})
 }
